moved embedding_matrix from children to the parent (AutoregressiveDecoder)
