+++ neuralmonkey/decoders/autoregressive.py
@@ -13,8 +13,9 @@ import tensorflow as tf
 
 from neuralmonkey.dataset import Dataset
 from neuralmonkey.decorators import tensor
-from neuralmonkey.logging import log
 from neuralmonkey.model.model_part import ModelPart, FeedDict, InitializerSpecs
+from neuralmonkey.logging import log, warn
+from neuralmonkey.model.sequence import EmbeddedSequence
 from neuralmonkey.nn.utils import dropout
 from neuralmonkey.tf_utils import get_variable
 from neuralmonkey.vocabulary import Vocabulary, START_TOKEN
@@ -69,6 +70,8 @@ class AutoregressiveDecoder(ModelPart):
                  data_id: str,
                  max_output_len: int,
                  dropout_keep_prob: float = 1.0,
+                 embedding_size: int = None,
+                 embeddings_source: EmbeddedSequence = None,
                  label_smoothing: float = None,
                  save_checkpoint: str = None,
                  load_checkpoint: str = None,
@@ -82,6 +85,9 @@ class AutoregressiveDecoder(ModelPart):
             data_id: Target data series.
             max_output_len: Maximum length of an output sequence.
             dropout_keep_prob: Probability of keeping a value during dropout.
+            embedding_size: Size of embedding vectors for target words.
+            embeddings_source: Embedded sequence to take embeddings from.
+            label_smoothing: Label smoothing parameter.
         """
         ModelPart.__init__(self, name, save_checkpoint, load_checkpoint,
                            initializers)
@@ -92,6 +98,8 @@ class AutoregressiveDecoder(ModelPart):
         self.data_id = data_id
         self.max_output_len = max_output_len
         self.dropout_keep_prob = dropout_keep_prob
+        self.embedding_size = embedding_size
+        self.embeddings_source = embeddings_source
         self.label_smoothing = label_smoothing
 
         # check the values of the parameters (max_output_len, ...)
@@ -102,6 +110,20 @@ class AutoregressiveDecoder(ModelPart):
         if dropout_keep_prob < 0.0 or dropout_keep_prob > 1.0:
             raise ValueError("Dropout keep probability must be"
                              "a real number in the interval [0,1].")
+
+        if self.embedding_size is None and self.embeddings_source is None:
+            raise ValueError("You must specify either embedding size or the "
+                             "embedded sequence from which to reuse the "
+                             "embeddings (e.g. set either 'embedding_size' or "
+                             " 'embeddings_source' parameter)")
+
+        if self.embeddings_source is not None:
+            if self.embedding_size is not None:
+                warn("Overriding the embedding_size parameter with the"
+                     " size of the reused embeddings from the encoder.")
+
+            self.embedding_size = (
+                self.embeddings_source.embedding_matrix.get_shape()[1].value)
 
         with self.use_scope():
             self.train_mode = tf.placeholder(tf.bool, [], "train_mode")
@@ -131,6 +153,21 @@ class AutoregressiveDecoder(ModelPart):
             return get_variable(
                 "logit_bias", [len(self.vocabulary)],
                 initializer=tf.zeros_initializer())
+
+    @tensor
+    def embedding_matrix(self) -> tf.Variable:
+        """Variables and operations for embedding of input words.
+
+        If we are reusing word embeddings, this function takes the embedding
+        matrix from the first encoder
+        """
+        if self.embeddings_source is not None:
+            return self.embeddings_source.embedding_matrix
+
+        return tf.get_variable(
+            name="word_embeddings",
+            shape=[len(self.vocabulary), self.embedding_size],
+            initializer=tf.glorot_uniform_initializer())
 
     def get_logits(self, state: tf.Tensor) -> tf.Tensor:
         """Project the decoder's output layer to logits over the vocabulary."""
